try new commit return type

+++ string.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int my_strlen(char *s)
+size_t my_strlen(char *s)
 {
    char *p = s;
 
@@ -21,7 +21,7 @@ int main(void)
 
    for (i = 0; i < 2; ++i)
       
-   printf("string lenght of %s = %d\n", s[i], my_strlen(s[i]));
+   printf("string lenght of %s = %lu\n", s[i], my_strlen(s[i]));
 
    return 0;
 }
